Add upstream release tracking
This is untested, so it's not really ready for PR. The idea is that if
mozilla-services/syncstorage-rs issues a new release, the docker image
should be rebuild with the new tag.

There is still more work to be done. Right now the Dockerfile checks out
a specific commit (1dc421474df2a4ee973d529534cd0eb7189d58ce) and builds
it. We would need to override that and tell it to use the release tag
from this.